Change iterations limit types for zero.

+++ cli/src/main.rs
@@ -1,3 +1,4 @@
+#![recursion_limit = "256"]
 #![allow(clippy::too_many_arguments)]
 #![allow(clippy::cognitive_complexity)]
 
@@ -141,7 +142,7 @@ where
         let confidence = InConfidence {
           minimax_depth: (8.0 * complexity).round() as u32,
           uct_iterations: (100_000.0 * complexity).round() as u32,
-          zero_iterations: (1_000.0 * complexity).round() as usize,
+          zero_iterations: (1_000.0 * complexity).round() as u32,
         };
         let analysis = futures::executor::block_on(state.oppai.analyze(
           &mut state.rng,
